FIRST COMMIT
WARNING: This code DOES NOT WORK in current versions of Thunderbird and may break things. YOU HVE BEEN WARNED.

Code is in need of some attention, lots of it refers to 'rainmaker' which was a prvious plugin on the same lines, which was repurposed for this.

The code does not pass Mozilla quality checks at present, so cannot be listed on AMO.